Added CNC Panel and numeric keypad popup support to manually control the cnc from the ui for configuration
Added move slots to scanner.py to move cnc to center or along the defined path. Added properties to know when the scanner is working or when a scan is progress.
Added a new CNC panel to access those elements via a new stack layout in LeftPane.qml. Circle path now stores its intrinsic parameters.
Added an estimation of the time it would take for the cnc to move (though it tends to underestimate for now)